tests: rework with webpack
