+++ etc/data/one_paragraph.json
@@ -1,3 +1,3 @@
-window.define(function () {
-  return "<p>Viber has <strong>confirmed</strong> a situation earlier this morning in which Viber appeared to have been <small>hacked</small> by the <code>Syrian Electronic Army</code> (a pro-government group of computer hackers aligned with Syrian President Bashar al-Assad).</p>";
-  } );
+{
+  "html": "<p>Viber has <strong>confirmed</strong> a situation earlier this morning in which Viber appeared to have been <small>hacked</small> by the <code>Syrian Electronic Army</code> (a pro-government group of computer hackers aligned with Syrian President Bashar al-Assad).</p>"
+}
