+++ etc/data/simple.json
@@ -1,3 +1,3 @@
-window.define(function () {
-  return "<p>Lorem <strong>ipsum <small>dolor</small> <a><span>amet</span></a></strong> sid.</p><p>Lorem <strong>ipsum <small>dolor</small> <a><span>amet</span></a></strong> sid.</p>";
-  } );
+{
+  "html": "<p>Lorem <strong>ipsum <small>dolor</small> <a><span>amet</span></a></strong> sid.</p><p>Lorem <strong>ipsum <small>dolor</small> <a><span>amet</span></a></strong> sid.</p>"
+}
