+++ etc/data/spaces.json
@@ -1,3 +1,3 @@
-window.define(function() {
-  return "<p>a    b</p><p>A paragraph with    more    whitespace!</p>";
-});
+{
+  "html": "<p>a    b</p><p>A paragraph with    more    whitespace!</p>"
+}
