+++ package.json
@@ -18,7 +18,7 @@
     "url": "https://github.com/dossier/html-highlighter.git/issues"
   },
   "peerDependencies": {
-    "jquery": "~2.2.4"
+    "jquery": "~3.1.0"
   },
   "dependencies": {},
   "devDependencies": {
@@ -26,7 +26,9 @@
     "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.5",
     "babel-preset-es2015": "^6.13.2",
+    "chai": "^3.5.0",
     "documentation-loader": "^0.1.1-beta",
+    "dom-loader": "^1.0.4",
     "eslint": "^3.3.1",
     "eslint-loader": "^1.5.0",
     "file-loader": "^0.9.0",
@@ -35,13 +37,23 @@
     "grunt-contrib-connect": "~1.0.2",
     "grunt-contrib-watch": "~1.0.0",
     "grunt-saucelabs": "~9.0.0",
+    "html-loader": "^0.4.3",
+    "html-webpack-plugin": "^2.22.0",
     "http-server": "^0.9.0",
+    "jsdom": "^9.4.2",
+    "jsdom-global": "^2.1.0",
+    "json-loader": "^0.5.4",
+    "mocha": "^3.0.2",
+    "mocha-loader": "^0.7.1",
+    "mocha-webpack": "^0.6.0",
+    "source-map-support": "^0.4.2",
     "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.15.0"
+    "webpack-dev-server": "^1.15.0",
+    "webpack-node-externals": "^1.3.3"
   },
   "scripts": {
     "prebuild": "npm install",
     "build": "webpack --progress --color",
-    "test": "grunt test"
+    "test": "mocha-webpack --webpack-config webpack.config-test.js \"test/start.js\" -c #grunt test"
   }
 }
