--- test/data
@@ -1,1 +1,0 @@
-../etc/data
