+++ test/html/templates.html
@@ -1,0 +1,12 @@
+<script type="text/hh-template" id="entity-row">
+ <li>
+ <input type="checkbox" data-hh-scope="enable">
+ <span class="name" data-hh-scope="name"></span>:
+                                  <span class="count" data-hh-scope="count"></span>
+ <span class="remove glyphicon glyphicon-remove" data-hh-scope="remove"></span>
+ </li>
+</script>
+
+<script type="text/hh-template" id="entity-empty">
+ <p class="empty">No query sets have been created yet</p>
+</script>
