+++ test/html/ui.html
@@ -1,0 +1,16 @@
+<div id="container" style="display: none">
+  <div id="document"></div>
+  <div id="widget" data-hh-scope="widget">
+    <div id="entities" data-hh-scope="entities"></div>
+    <span data-hh-scope="expand"></span>
+    <span>
+      <span id="stats-current" data-hh-scope="stats-current">-</span>
+      /
+      <span id="stats-total" data-hh-scope="stats-total">-</span>
+    </span>
+    <div role="group" aria-label="Navigation">
+      <button type="button" data-hh-scope="button-next"></button>
+      <button type="button" data-hh-scope="button-prev"></button>
+    </div>
+  </div>
+</div>
