--- test/index.html
@@ -1,98 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Mocha Spec Runner</title>
-
-  <link rel="stylesheet" href="lib/mocha-2.2.1/mocha.css">
-
-  <!-- Template definition -->
-  <script type="text/hh-template" id="entity-row">
-    <li>
-      <input type="checkbox" data-hh-scope="enable">
-      <span class="name" data-hh-scope="name"></span>:
-      <span class="count" data-hh-scope="count"></span>
-      <span class="remove glyphicon glyphicon-remove" data-hh-scope="remove"></span>
-    </li>
-  </script>
-
-  <script type="text/hh-template" id="entity-empty">
-   <p class="empty">No query sets have been created yet</p>
-  </script>
-
-  <!-- Include dependencies -->
-  <!-- Mocha and Chai are NOT loaded by RequireJs so we may benefit from the
-       convenience of global scope references. -->
-  <script src="lib/mocha-2.2.1/mocha.js"></script>
-  <script src="lib/chai-2.1.0.js"></script>
-	<script src="lib/requirejs-2.1.16/requirejs.js"></script>
-
-  <!-- Set up mocha  -->
-  <script type="text/javascript">mocha.setup('bdd')</script>
-  <script>
-   require.config( {
-     paths: {
-       'jquery': 'lib/jquery-2.1.3/jquery',
-       'SortingCommon': 'lib/sorting-common/sorting_common',
-       'data1': 'data/viber_attacked_by_syrian_electronic_army-cropped.json?',
-       'data2': 'data/one_paragraph-ampersand_nonexistent.json?',
-       'data3': 'data/one_paragraph-ampersand.json?',
-       'data4': 'data/one_paragraph-ampersand_escaped.json?',
-       'html_highlighter': 'src/html_highlighter'
-     }
-   } );
-
-   require([ 'spec/html_highlighter' ], function (spec) {
-     var runner = mocha.run(),
-         failedTests = [];
-
-     runner.on('end', function () {
-       window.mochaResults = runner.stats;
-       window.mochaResults.reports = failedTests;
-     });
-
-     runner.on('fail', function (test, err) {
-       var flattenTitles = function(test) {
-         var titles = [];
-         while (test.parent.title){
-           titles.push(test.parent.title);
-           test = test.parent;
-         }
-         return titles.reverse();
-       };
-
-       failedTests.push( {
-         name: test.title,
-         result: false,
-         message: err.message,
-         stack: err.stack,
-         titles: flattenTitles(test)
-       } );
-     } );
-   } );
-  </script>
-</head>
-
-<body>
-  <div id="mocha"><p><a href=".">Index</a></p></div>
-  <div id="messages"></div>
-  <div id="fixtures"></div>
-
-  <div id="container" style="display: none">
-    <div id="document"></div>
-    <div id="widget" data-hh-scope="widget">
-      <div id="entities" data-hh-scope="entities"></div>
-      <span data-hh-scope="expand"></span>
-      <span>
-        <span id="stats-current" data-hh-scope="stats-current">-</span>
-        /
-        <span id="stats-total" data-hh-scope="stats-total">-</span>
-      </span>
-      <div role="group" aria-label="Navigation">
-        <button type="button" data-hh-scope="button-next"></button>
-        <button type="button" data-hh-scope="button-prev"></button>
-      </div>
-    </div>
-  </div>
-</body>
-</html>
