--- test/src
@@ -1,1 +1,0 @@
-../dist
