+++ test/start.js
@@ -1,0 +1,2 @@
+const context = require.context('mocha!.', true, /_test.js$/);
+context.keys().forEach(context);
